@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.18.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,20 +26,20 @@ dependencies:
   flutter_localizations:
       sdk: flutter
   intl: ^0.17.0
-  uuid: 3.0.4
-  http: 0.13.3
-  json_annotation: 4.1.0
+  uuid: 3.0.7
+  http: 0.13.5
+  json_annotation: 4.7.0
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: 2.0.6
-  json_serializable: 5.0.0
+  build_runner: 2.3.2
+  json_serializable: 6.5.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
