@@ -1,8 +1,8 @@
 import 'dart:convert';
 import 'dart:developer';
-import 'package:path_of_market/api/urlBuilder.dart';
-import 'package:path_of_market/enums/categoryTypeEnum.dart';
-import 'package:path_of_market/models/categoryModels.dart';
+import 'package:path_of_market/api/ur_builder.dart';
+import 'package:path_of_market/enums/category_type.dart';
+import 'package:path_of_market/models/category_models.dart';
 import 'package:http/http.dart' as http;
 
 class MainApiService {
@@ -10,7 +10,7 @@ class MainApiService {
   final UrlBuilder urlBuilder = UrlBuilder();
   String _url = '';
 
-  http.Response response;
+  http.Response? response;
 
   Future<List<Item>> getItemList(CategoryType catType) async {
 
@@ -28,19 +28,19 @@ class MainApiService {
 
     response = await http.get(Uri.parse(_url));
 
-    if (response.statusCode == 200) {
+    if (response!.statusCode == 200) {
       log('${catType.name} request was successfull.');
 
       if (catType == CategoryType.currency) {
-        resultList.addAll(_decodeCurrencyItems(json.decode(response.body)));
+        resultList.addAll(_decodeCurrencyItems(json.decode(response!.body)));
       } else {
-        resultList.addAll(_decodeRegularItems(json.decode(response.body)));
+        resultList.addAll(_decodeRegularItems(json.decode(response!.body)));
       }
     } else {
       log('Failed to load ${catType.name} category. ' +
-          response.statusCode.toString() +
+          response!.statusCode.toString() +
           ' ' +
-          response.reasonPhrase);
+          response!.reasonPhrase!);
     }
   }
 
