@@ -1,5 +1,5 @@
-import 'package:path_of_market/enums/categoryTypeEnum.dart';
-import 'package:path_of_market/utils/apiConstants.dart';
+import 'package:path_of_market/enums/category_type.dart';
+import 'package:path_of_market/utils/api_constants.dart';
 
 class UrlBuilder {
 
@@ -8,7 +8,7 @@ class UrlBuilder {
 
   String buildUrl(bool isCurrencyCategory, CategoryType categoryType) {
     String defaultString = MAIN_URL;
-    String catType = categoryType.name;
+    String catType = categoryType.name ?? '';
 
     defaultString += (isCurrencyCategory ? CATEGORY_CURRENCY_PATH : CATEGORY_ITEM_PATH);
     defaultString += OR_DELIMETER;
