@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:path_of_market/models/categoryModels.dart';
-import 'package:path_of_market/ui/itemList.dart';
-import 'package:path_of_market/enums/categoryTypeEnum.dart';
+import 'package:path_of_market/models/category_models.dart';
+import 'package:path_of_market/ui/item_list.dart';
+import 'package:path_of_market/enums/category_type.dart';
 import 'package:uuid/uuid.dart';
 
 class CategoryTile extends StatefulWidget {
   final Category _category;
-  final void Function(Category) onCategoryTap;
+  final void Function(Category)? onCategoryTap;
 
   CategoryTile(this._category, { this.onCategoryTap });
 
@@ -23,12 +23,12 @@ class _CategoryTileState extends State<CategoryTile> {
     return ListTile(
       key: Key(uuid.v1()),
       leading: widget._category.icon,
-      title: Text(widget._category.categoryType.name),
+      title: Text(widget._category.categoryType.name ?? ''),
       onTap: () {
         if (widget.onCategoryTap == null) {
           _defaultCategoryTap();
         } else {
-          widget.onCategoryTap(widget._category);
+          widget.onCategoryTap!(widget._category);
         }
       },
     );
