@@ -2,6 +2,6 @@ const MAIN_URL = "https://poe.ninja/api/data/";
 const CATEGORY_CURRENCY_PATH = "CurrencyOverview";
 const CATEGORY_ITEM_PATH = "ItemOverview";
 
-const CURRENT_LEAGUE_NAME = "Sanctum";
+const CURRENT_LEAGUE_NAME = "Kalandra";
 const LANGUAGE_PATH_RU = "ru";
 const LANGUAGE_PATH_ENG = "en";
